requirements.txt
tensorflow==1.15.0
keras
flask
gevent
requests
Pillow